refactor: gameUtils 기능을 class Game으로 통합
